Разделил по файлам, создал структуру проекта
